+++ changelog.md
@@ -26,6 +26,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityTablePanel bug fixed, summary panel toggle control displayed only if a summary panel is available.
 - EntityApplicationPanel.createMainMenuControls() and related methods now Optional based.
+- EntityApplicationPanel, control for opening log folder added to help menu, related refactoring.
 ### is.codion.swing.framework.server.monitor
 - ServerMonitorPanel, ClientUserMonitorPanel, shutdown and disconnect confirmation changed from OK/Cancel to Yes/No.